+++ lib/app.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+
+class App extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Material App',
+      home: Scaffold(
+        appBar: AppBar(
+          title: const Text('Material App Bar'),
+        ),
+        body: Center(
+          child: Container(
+            child: const Text('Hello World'),
+          ),
+        ),
+      ),
+    );
+  }
+}