+++ lib/design/breakpoints.dart
@@ -1,0 +1,7 @@
+// ignore_for_file: public_member_api_docs
+mixin SkeletonBreakPoint {
+  static const double s = 599;
+  static const double m = 899;
+  static const double l = 1199;
+  static const double xl = 1200;
+}