+++ lib/main.dart
@@ -1,22 +1,5 @@
 import 'package:flutter/material.dart';
 
-void main() => runApp(App());
+import 'app.dart';
 
-class App extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Material App',
-      home: Scaffold(
-        appBar: AppBar(
-          title: Text('Material App Bar'),
-        ),
-        body: Center(
-          child: Container(
-            child: Text('Hello World'),
-          ),
-        ),
-      ),
-    );
-  }
-}
+void main() => runApp(App());